backend/requirements.txt
Django==4.2
djangorestframework==3.14
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
django-filter==25.1
python-decouple==3.8
